fix(sdk): add ensureResolvedProtocolVersion, remove circular dep
Extracts version-resolution logic into a dedicated `resolveFhevmVersions-p.ts` module to break a circular dependency introduced in the previous commit. Adds `ensureResolvedProtocolVersion` as a guard used by client decorators before any encrypt/decrypt operation, and slims down `CoreFhevm-p` accordingly.
