+++ sdk/js-sdk/src/core/clients/decorators/decrypt-p.ts
@@ -1,20 +1,14 @@
 import type { FhevmBase } from '../../types/coreFhevmClient.js';
 import type { FhevmChain } from '../../types/fhevmChain.js';
-import {
-  asFhevmClientWith,
-  resolveFhevmProtocolVersion,
-  resolveFhevmTkmsVersion,
-  setResolvedProtocolVersion,
-  setResolvedTkmsVersion,
-} from '../../runtime/CoreFhevm-p.js';
+import { asFhevmClientWith, setResolvedTkmsVersion } from '../../runtime/CoreFhevm-p.js';
+import { ensureResolvedProtocolVersion, resolveFhevmTkmsVersion } from '../../runtime/resolveFhevmVersions-p.js';
 
 ////////////////////////////////////////////////////////////////////////////////
 
 export async function _initDecrypt(fhevm: FhevmBase<FhevmChain>): Promise<void> {
   const f = asFhevmClientWith(fhevm, 'decrypt');
 
-  const protocolVersion = await resolveFhevmProtocolVersion(f);
-  setResolvedProtocolVersion(f, protocolVersion);
+  await ensureResolvedProtocolVersion(f);
 
   const tkmsVersion = await resolveFhevmTkmsVersion(f);
