+++ sdk/js-sdk/src/core/clients/decorators/encrypt-p.ts
@@ -1,21 +1,15 @@
 import type { FhevmBase } from '../../types/coreFhevmClient.js';
 import type { FhevmChain } from '../../types/fhevmChain.js';
-import {
-  asFhevmClientWith,
-  resolveFhevmProtocolVersion,
-  resolveFhevmTfheVersion,
-  setResolvedProtocolVersion,
-  setResolvedTfheVersion,
-} from '../../runtime/CoreFhevm-p.js';
+import { asFhevmClientWith, setResolvedTfheVersion } from '../../runtime/CoreFhevm-p.js';
 import { fetchFheEncryptionKeyBytes } from '../../key/fetchFheEncryptionKeyBytes.js';
+import { ensureResolvedProtocolVersion, resolveFhevmTfheVersion } from '../../runtime/resolveFhevmVersions-p.js';
 
 ////////////////////////////////////////////////////////////////////////////////
 
 export async function _initEncrypt(fhevm: FhevmBase<FhevmChain>): Promise<void> {
   const f = asFhevmClientWith(fhevm, 'encrypt');
 
-  const protocolVersion = await resolveFhevmProtocolVersion(f);
-  setResolvedProtocolVersion(f, protocolVersion);
+  await ensureResolvedProtocolVersion(fhevm);
 
   const tfheVersion = await resolveFhevmTfheVersion(f);
